requirements.txt
pyttsx3
speechrecognition
wikipedia
python-dotenv
playsound
requests
mutagen
pipwin
win10toast
prettytable